backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Neo4j
neo4j==5.14.1

# Redis
redis==5.0.1
aioredis==2.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP client for OpenRouter
httpx==0.25.2
openai==1.3.7

# Data validation
pydantic[email]==2.5.0

# Environment variables
python-dotenv==1.0.0

# Logging and monitoring
structlog==23.2.0

# Date and time
python-dateutil==2.8.2

# Cryptography
cryptography==41.0.7

# Background tasks
celery==5.3.4

# Email
emails==0.6

# File handling
python-magic==0.4.27

# Testing (development)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Code quality (development)
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1